feat: automatically update players list
The online players list on the webinterface now updates automatically
when players connect or disconnect. This is done by using the new events
feature in the NetworkManager, to give us a callback when a client
connects/disconnects. That an event occurred is registered and handled
by an async handler in the node module. Then we can expose a route that
provides an SSE stream that the frontend can listen on and update the UI
to.