feat: reimplement password reset
This refactors and reimplements the password reset via email
functionality. By entering your email in the login field in the UI and
hitting "Reset Password", the backend will generate a link with a
password reset token and send it in an email to the user. Via the link,
the user then has the ability to enter a new password for their account.
This only works if the user uses email/password auth. The link sent to
the user is valid for 5 hours.